Do not try to convert files not matching the expected input format
